requirements.txt
Jinja2==2.7.1
MarkupSafe==0.18
PyYAML==3.10
argparse==1.2.1
distribute==0.6.24
html5lib==0.95
pisa==3.0.33
pyPdf==1.13
reportlab==2.7
wsgiref==0.1.2
